1.1-alpha1 Complete reconstruction No more are the simple "levels of randomization"... I have completely reconstructed the randomizer by breaking out all of the possible ways you can randomize this game.  If you no longer want monster stats/patterns randomized, you don't have to any more.
The statistics randomization has been reconstructed as well.  There is now a maximum statistic, with 35/45/50 randomized numbers reaching that maximized statistic.  (This is basically how it's done in Dragon Warrior 1 Randomizer)

Enemy resistances were reconstructed as well; they are completely randomized in ludicrous randomness, and adjusted +/- in lesser levels.

Enemy patterns were reconstructed as well.  In ludicrous randomness, all monsters' patterns are completely randomized.  In lesser levels, they can either be randomized, randomized based on how tough the monster is supposed to be, or possibly maintain their original pattern!

I removed the chance to remove a monster in the monster zones.  Each zone will now have the full six monsters in it.  I think this might increase the encounter rate a little, but I'm not too worried about that.

Forced six items in ALL stores.  But items could be removed if duplicates are drawn from the possible selection.

The golden key is no longer available for purchase.  Too often you were able to skip the Dragon's Horn; it should be possible, but not almost every game!